fix/修复Chat模块Controller层抽象的路径分配和命名
